# public/swagger/api.json
{
    "openapi": "3.0.0",
    "servers": [
        {
            "url": "http://localhost:2222/",
            "description": "OpenApi host"
        }
    ],
    "paths": {
...
        "\/api\/categories\/{id}": {
            "delete": {
                "tags": [
                    "Category"
                ],
                "description": "delete category",
                "operationId": "App\\Http\\Controllers\\Api\\CategoryController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete category success"
                    }
                }
            }
        },
...
        "\/api\/courses\/{id}": {
            "get": {
                "tags": [
                    "Course"
                ],
                "description": "show course detail",
                "operationId": "App\\Http\\Controllers\\Api\\CourseController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "course id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response course success",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "data": {
                                        "id": 1,
                                        "name": "sr7Mlx8itY",
                                        "description": "Wd00fJflFQ",
                                        "is_active": 0,
                                        "category_id": 4,
                                        "created_at": "2020-12-12",
                                        "updated_at": "2020-12-12"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invalid course id"
                    }
                }
            },
            "post": {
                "tags": [
                    "Course"
                ],
                "description": "Update course",
                "operationId": "App\\Http\\Controllers\\Api\\CourseController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "course id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Store course success"
                    },
                    "422": {
                        "description": "Missing or invalid data"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Course"
                ],
                "description": "delete course",
                "operationId": "App\\Http\\Controllers\\Api\\CourseController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "course id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete course success"
                    },
                    "404": {
                        "description": "Invalid course id"
                    }
                }
            }
        },
...
        "\/api\/subjects\/{id}": {
            "get": {
                "tags": [
                    "Subject"
                ],
                "description": "show subject detail",
                "operationId": "App\\Http\\Controllers\\Api\\SubjectController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "subject id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response subject success",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "data": {
                                        "id": 10,
                                        "name": "rxe8KR5xqc",
                                        "description": "RzmagvNtXV",
                                        "is_active": 1,
                                        "created_at": null,
                                        "updated_at": null
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Missing or invalid data"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Subject"
                ],
                "description": "delete subject",
                "operationId": "App\\Http\\Controllers\\Api\\SubjectController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "subject id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete subject success"
                    }
                }
            }
        },
...
        "\/api\/tasks?name={name}&perPage={perPage}&page={page}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "description": "Get task",
                "operationId": "App\\Http\\Controllers\\Api\\TaskController::index",
                "parameters": [
                    {
                        "name": "name",
                        "default": " ",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "perPage",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "path",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response tasks success",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "data": {
                                        "data": {
                                            "id": "1",
                                            "name": "task1",
                                            "description": "dCXatgvWk2",
                                            "content": "rPeYC62M5k",
                                            "deadline": "2020-10-10",
                                            "is_active": 1,
                                            "created_at": "2020-10-10",
                                            "updated_at": "2020-10-10"
                                        },
                                        "current_page": 1,
                                        "first_page_url": "http:\/\/localhost:2222\/api\/tasks?page=1",
                                        "from": 1,
                                        "last_page": 1,
                                        "last_page_url": "http:\/\/localhost:2222\/api\/tasks?page=1",
                                        "next_page_url": null,
                                        "path": "http:\/\/localhost:2222\/api\/tasks",
                                        "per_page": 15,
                                        "prev_page_url": null,
                                        "to": 10,
                                        "total": 10
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "\/api\/tasks": {
            "post": {
                "tags": [
                    "Task"
                ],
                "description": "Store task",
                "operationId": "App\\Http\\Controllers\\Api\\TaskController::store",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "properties": {
                                    "task": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "content": {
                                                "type": "string"
                                            },
                                            "deadline": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "is_active": {
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "subject_id": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "user_id": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Store task success"
                    },
                    "422": {
                        "description": "Missing or invalid data"
                    }
                }
            }
        },
        "\/api\/tasks\/{id}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "description": "show task detail",
                "operationId": "App\\Http\\Controllers\\Api\\TaskController::show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response task success"
                    },
                    "422": {
                        "description": "Missing or invalid data"
                    }
                }
            },
            "put": {
                "tags": [
                    "Task"
                ],
                "description": "Update task",
                "operationId": "App\\Http\\Controllers\\Api\\TaskController::update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application\/json": {
                            "schema": {
                                "properties": {
                                    "task": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "content": {
                                                "type": "string"
                                            },
                                            "deadline": {
                                                "type": "string",
                                                "format": "date"
                                            },
                                            "is_active": {
                                                "type": "boolean"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "subject_id": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    },
                                    "user_id": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update task success"
                    },
                    "422": {
                        "description": "Missing or invalid data"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Task"
                ],
                "description": "delete task",
                "operationId": "App\\Http\\Controllers\\Api\\TaskController::destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "delete task success"
                    }
                }
            }
        },
        "\/api\/tasks\/subjects\/{id}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "description": "Get subjects in task",
                "operationId": "App\\Http\\Controllers\\Api\\TaskController::getSubjectOfTask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "response success",
                        "content": {
                            "application\/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "examples": {
                                    "data": {
                                        "id": 3,
                                        "name": "Yytfgr9MMY",
                                        "description": "rfzJwaZg77",
                                        "is_active": 0,
                                        "created_at": null,
                                        "updated_at": null,
                                        "pivot": {
                                            "task_id": 1,
                                            "subject_id": 3
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}